Possibly fix the SpriteBatch memory leak
Fixed by writing an optimisation in SpriteBatch to not create new VertexBuffers in each Draw() update, but to modify an existing Vertex Buffer with an initial capacity of 1024 and a growth factor of 2.

Fixed VertexBuffer usage variable not being used in the ctor. Thus also fixing
::SetData() throwing exceptions.

Finally, also fixed the post build process.